@@ -18,15 +18,15 @@
     "@nuxtjs/auth-next": "^5.0.0-1648802546.c9880dc",
     "@nuxtjs/axios": "^5.13.6",
     "@nuxtjs/composition-api": "^0.32.0",
-    "@nuxtjs/i18n": "^7.2.2",
+    "@nuxtjs/i18n": "^7.3.0",
     "@nuxtjs/svg-sprite": "^0.5.2",
     "@types/body-scroll-lock": "^3.1.0",
     "@vue-a11y/announcer": "^2.1.0",
     "@vue/apollo-composable": "^4.0.0-alpha.19",
-    "@vuelidate/core": "^2.0.0-alpha.42",
-    "@vuelidate/validators": "^2.0.0-alpha.30",
+    "@vuelidate/core": "^2.0.0-alpha.44",
+    "@vuelidate/validators": "^2.0.0-alpha.31",
     "body-scroll-lock": "^3.1.5",
-    "core-js": "^3.23.5",
+    "core-js": "^3.25.0",
     "graphql-tag": "^2.12.6",
     "ipx": "^0.9.10",
     "nuxt": "^2.15.8",
@@ -49,21 +49,21 @@
     "@types/smoothscroll-polyfill": "^0.3.1",
     "@types/uuid": "^8.3.4",
     "apollo-boost": "^0.4.9",
-    "eslint": "^8.20.0",
+    "eslint": "^8.23.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-nuxt": "^3.2.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-vue": "^9.2.0",
-    "node-fetch": "^3.2.9",
+    "eslint-plugin-vue": "^9.4.0",
+    "node-fetch": "^3.2.10",
     "postcss-html": "^1.5.0",
     "postcss-mixins": "^9.0.3",
-    "postcss-preset-env": "^7.7.2",
+    "postcss-preset-env": "^7.8.0",
     "prettier": "^2.7.1",
     "prettier-stylelint": "^0.4.2",
-    "stylelint": "^14.9.1",
+    "stylelint": "^14.11.0",
     "stylelint-config-prettier": "^9.0.3",
     "stylelint-config-recommended-vue": "^1.4.0",
-    "stylelint-config-standard": "^26.0.0",
+    "stylelint-config-standard": "^28.0.0",
     "stylelint-prettier": "^2.0.0"
   }
 }
