@@ -1,65 +1,69 @@
 <template>
   <section :class="$style['app-stores']" aria-labelledby="app-stores-title">
-    <div :class="$style.wrapper">
-      <div :class="$style['image-wrapper']">
-        <nuxt-picture
-          :class="$style.mood"
-          src="/images/arcarde.png"
-          width="500"
-          height="560"
-          alt=""
-          loading="lazy"
-          sizes="sm:100vw lg:75vw xl:33v xl:500px xxl:1000px"
-          preset="base"
-          format="avif"
-        />
-
-        <button :class="$style['btn-action']" @click="action">
+    <center-wrapper size="lg">
+      <div :class="$style.wrapper">
+        <div :class="$style['image-wrapper']">
           <nuxt-picture
-            src="/images/arcarde-button.png"
-            width="36"
-            height="34"
-            alt="Play"
-            sizes="xs:36 sm:72"
+            :class="$style.mood"
+            src="/images/arcarde.png"
+            width="500"
+            height="560"
+            alt=""
+            loading="lazy"
+            sizes="sm:100vw lg:75vw xl:33v xl:500px xxl:1000px"
             preset="base"
             format="avif"
           />
-        </button>
+
+          <button :class="$style['btn-action']" @click="action">
+            <nuxt-picture
+              src="/images/arcarde-button.png"
+              width="36"
+              height="34"
+              alt="Play"
+              sizes="xs:36 sm:72"
+              preset="base"
+              format="avif"
+            />
+          </button>
+        </div>
+        <div :class="$style.text">
+          <h1 id="app-stores-title">Loesje in de App Stores</h1>
+          <p>
+            De posters van Loesje ken je vooral van op prullenbakken in de stad
+            of een elektriciteitskastje bij je om de hoek, maar Loesje wil de
+            wereld ook graag digitaal mooier maken. Dagelijks komt er al een
+            Tweet en Facebook-post voorbij vliegen, maar vanaf nu is er meer!
+          </p>
+          <p>
+            Nu is er de officiele Loesje-app. In deze app kan je uren
+            rondstruinen op zoek de poster die jij het tofst vindt. De app is te
+            downloaden via de
+            <a
+              href="https://itunes.apple.com/nl/app/loesje-posters/id910472463?l=nl&amp;ls=1&amp;mt=8"
+              target="_blank"
+              rel="noopener noreferrer"
+              >Appstore voor iOS </a
+            >en via de
+            <a
+              href="https://play.google.com/store/apps/details?id=nl.loesje&amp;hl=nl"
+              target="_blank"
+              rel="noopener noreferrer"
+              >Google Play Store</a
+            >.
+          </p>
+        </div>
       </div>
-      <div :class="$style.text">
-        <h1 id="app-stores-title">Loesje in de App Stores</h1>
-        <p>
-          De posters van Loesje ken je vooral van op prullenbakken in de stad of
-          een elektriciteitskastje bij je om de hoek, maar Loesje wil de wereld
-          ook graag digitaal mooier maken. Dagelijks komt er al een Tweet en
-          Facebook-post voorbij vliegen, maar vanaf nu is er meer!
-        </p>
-        <p>
-          Nu is er de officiele Loesje-app. In deze app kan je uren rondstruinen
-          op zoek de poster die jij het tofst vindt. De app is te downloaden via
-          de
-          <a
-            href="https://itunes.apple.com/nl/app/loesje-posters/id910472463?l=nl&amp;ls=1&amp;mt=8"
-            target="_blank"
-            rel="noopener noreferrer"
-            >Appstore voor iOS </a
-          >en via de
-          <a
-            href="https://play.google.com/store/apps/details?id=nl.loesje&amp;hl=nl"
-            target="_blank"
-            rel="noopener noreferrer"
-            >Google Play Store</a
-          >.
-        </p>
-      </div>
-    </div>
+    </center-wrapper>
   </section>
 </template>
 
 <script lang="ts">
 import { defineComponent } from '@nuxtjs/composition-api'
+import CenterWrapper from '../Wrappers/CenterWrapper.vue'
 
 export default defineComponent({
+  components: { CenterWrapper },
   setup() {
     let themeColor = '#000'
     const action = () => {
@@ -80,8 +84,6 @@ export default defineComponent({
 }
 
 .wrapper {
-  @mixin center var(--container-width-lg);
-
   position: relative;
   display: flex;
   flex-direction: column;
